+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -20,7 +20,7 @@ public class GenreController {
     }
 
     @GetMapping("/genres/{id}")
-    public Genre getGenreById(@PathVariable long id) {
+    public Genre getGenreById(@PathVariable Integer id) {
         return genreService.findById(id);
     }
 }
